requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.2
SQLAlchemy==2.0.43
python-dotenv==1.0.0
# Linux/macOS servers (Render)
gunicorn==23.0.0; sys_platform != "win32"
psycopg2-binary==2.9.10; sys_platform != "win32"
# Windows local dev
waitress==3.0.0; sys_platform == "win32"